Process arguments in the window.
This will eventually ported to other modules (such as the DSPs or the visualisers).
For now though, just do the window to establish an interface.

An argexception class has also been added to indicate an improperly formatted arguement.
This should trigger the main code to call a usage function to print out the usage
functions of all the modules that the visualiser uses (including the window), well that's
the idea anyway.

Also I have incremented the library version to 1:0:1. We have a new interface, however
this is a superset of the old interface and thus backwards compatable. I would like all
code, however, to switch to the new interface, so I have marked the old visualiserWin
constructor as depricated.